+++ R/github_dependents.R
@@ -28,47 +28,22 @@ github_dependents <- function(owner,
                               max_pages = 1000,
                               verbose = TRUE) {
     
-    #devoptera::args2vars(github_insights)  
+    # devoptera::args2vars(github_dependents)  
     
-    requireNamespace("rvest")
-    owner_repo <- NULL;
     messager("Searching for dependents of:",paste(owner,repo,sep="/"),
              v=verbose) 
-    url <- paste0("https://github.com/", owner,"/",repo, "/network/dependents")
-    #### Loop over the specified number of pages ####
-    all_dat <- list() 
-    for (i in seq_len(max_pages)) {
-        # Print a message indicating the URL being scraped
-        messager(paste0("+ Scraping page ",i,"."),v=verbose) 
-        # Retrieve the HTML content of the page
-        page <- rvest::read_html(url) 
-        box_rows <- rvest::html_elements(page,".Box-row")
-        dt <- (rvest::html_text2(box_rows)) |> 
-            stringr::str_split(" / |\n|[ ]", simplify = TRUE) |> 
-            data.table::data.table() |> 
-            `colnames<-`(c("owner","repo","stargazers_count","forks_count"))
-        dt[,owner_repo:=paste(owner,repo,sep="/")]
-        dt <- cbind(target=paste(owner,repo,sep="/"),dt) 
-        all_dat[[i]] <- dt 
-        #### Find the button for the next page ####
-        buttons <- page |> rvest::html_nodes(".paginate-container .btn")
-        next_buttons <- buttons[rvest::html_text(buttons)=="Next"]
-        #### Check if the button is disabled ####
-        is_disabled <- any(sapply(next_buttons, function(btn) {
-            btn_attr <- rvest::html_attr(btn, "disabled")
-            !is.na(btn_attr) && btn_attr == "disabled"
-        }))
-        #### If the button isn't disable, update the URL to scrape ####
-        if (isFALSE(is_disabled)) { 
-            url <- next_buttons |> rvest::html_attr("href") 
-        #### Otherwise, break the loop ####
-        } else {
-            break 
-        }
-    }
-    #### Bind data from all pages ####
-    all_dat <- data.table::rbindlist(all_dat, 
-                                     use.names = TRUE, idcol = "page")
+    
+    #### Method 1: JSON file ####
+    # URL <- paste0("https://github.com/", owner,"/",repo, "/dependency-graph/sbom")
+    # j <- jsonlite::fromJSON("~/Downloads/rworkflows_neurogenomics_b017b7a1aeda0026dda330b01cb798ddb5f1d264.json")
+    # j$packages
+    
+    #### Method 2: Webscraping ####
+    all_dat <- github_dependents_scrape(owner = owner,
+                                        repo = repo, 
+                                        token = token,
+                                        max_pages = max_pages,
+                                        verbose = verbose)
     #### Report ####
     messager("Found",formatC(nrow(all_dat),big.mark = ","),
              "dependents.",v=verbose)